BudgetWatchDogV2/BudgetBackend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cache2==0.2.2
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
kiwisolver==1.4.7
matplotlib==3.10.0
numpy==2.2.0
openai==1.58.1
packaging==24.2
pendulum==3.0.0
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
time-machine==2.16.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
